pyproject.toml
[tool.poetry]
name = "dataset_collctor"
version = "0.1.0"
description = ""
authors = ["daichitominaga <daichi.tominaga.1112@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.3"
click = "^8.1.3"
gitpython = "^3.1.29"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
transformers = "^4.27.4"
black = "^23.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
